backend/rag/requirements.txt
python-dotenv
scikit-learn
faiss-cpu
pypdf

langchain-core
langchain-community
langchain-gigachat
gigachat

fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0